+++ bigmler/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__version__ = '0.8.0'
+__version__ = '1.0'
 
 if __name__ == '__main__':
     import sys